+++ API/DTOs/LoginDto.cs
@@ -2,6 +2,6 @@
 
 public class LoginDto
 {
-    public int CardNum { get; set; }
+    public long CardNum { get; set; }
     public int PIN { get; set; }
 }
